moved hepshell.commands.Command to hepshell.Command
