--- hepshell/commands/check/__init__.py
@@ -1,18 +1,0 @@
-"""
-    check:   Used to check the status of various things
-        Usage:
-                check <thing>
-"""
-from .. import Command as C
-
-
-class Command(C):
-
-    def __init__(self, path=__file__, doc=__doc__):
-        super(Command, self).__init__(path, doc)
-
-    def run(self, args, variables):
-        self.__prepare(args, variables)
-        self.__text = "NOT IMPLEMENTED"
-
-        return True
