+++ hepshell/commands/dummy/__init__.py
@@ -10,11 +10,11 @@
         something: a weird parameter
 """
 import logging
-from .. import Command as C
+import hepshell
 
 LOG = logging.getLogger(__name__)
 
-class Command(C):
+class Command(hepshell.Command):
 
     def __init__(self, path=__file__, doc=__doc__):
         super(Command, self).__init__(path, doc)
